backend/pyproject.toml
[project]
name = "control-tower-backend"
version = "1.0.0"
description = "Unified Model, Tool & GenAI Control Tower – Backend API"
requires-python = ">=3.13"
license = { text = "Proprietary" }
authors = [{ name = "Control Tower Team" }]

dependencies = [
    # ── Web Framework ────────────────────────
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",

    # ── Database ─────────────────────────────
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "alembic>=1.14.0",
    "pgvector>=0.3.6",

    # ── Object Storage ───────────────────────
    "minio>=7.2.12",
    "boto3>=1.35.0",

    # ── Event Streaming ──────────────────────
    "aiokafka>=0.12.0",
    "confluent-kafka>=2.6.0",

    # ── Workflow Orchestration ───────────────
    "temporalio>=1.8.0",

    # ── Policy Engine ────────────────────────
    "opa-python-client>=1.3.0",

    # ── Observability ────────────────────────
    "opentelemetry-api>=1.29.0",
    "opentelemetry-sdk>=1.29.0",
    "opentelemetry-exporter-otlp>=1.29.0",
    "opentelemetry-instrumentation-fastapi>=0.50b0",
    "opentelemetry-instrumentation-sqlalchemy>=0.50b0",
    "opentelemetry-instrumentation-httpx>=0.50b0",
    "prometheus-client>=0.21.0",

    # ── LLM / AI ─────────────────────────────
    "openai>=1.58.0",
    "anthropic>=0.40.0",
    "langchain-core>=0.3.28",
    "langchain-openai>=0.3.0",
    "tiktoken>=0.8.0",

    # ── Evaluation & Tracing ─────────────────
    "arize-phoenix>=5.0.0",
    "langfuse>=2.56.0",
    "trulens-eval>=1.2.0",

    # ── Security & Red Teaming ───────────────
    "pyrit>=0.5.0",

    # ── Evidence & Reporting ─────────────────
    "reportlab>=4.2.0",
    "jinja2>=3.1.4",
    "markdown>=3.7",
    "pyyaml>=6.0.2",

    # ── Utilities ────────────────────────────
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.12",
    "python-dateutil>=2.9.0",
    "uuid7>=0.1.0",
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "factory-boy>=3.3.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-python-dateutil",
    "types-pyyaml",
    "types-markdown",
]

[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py313"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
